pyproject.toml
[project]
name = "gamelens-dashboard-service"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "flasgger>=0.9.7.1",
    "flask>=3.1.2",
    "gunicorn>=25.1.0",
    "psycopg[pool]>=3.3.2",
    "python-dotenv>=1.2.1",
]
